+++ package.json
@@ -11,15 +11,20 @@
     "lint:fix": "eslint --fix './src/**/*.{ts,tsx,js,jsx}'"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "@types/react": "~18.2.45",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "expo": "~50.0.13",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
+    "react-hook-form": "^7.51.1",
     "react-inlinesvg": "^4.1.3",
     "react-native": "0.73.5",
     "react-native-modal": "^13.0.1",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.29.0",
     "react-native-splash-screen": "^3.3.0",
     "react-native-svg": "^15.1.0",
     "styled-components": "^6.1.8"
