+++ src/screens/Register.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components/native';
+
+const Register = () => {
+  return null;
+};
+
+export default Register;
